move reputation module to latest rpc style
