+++ reputation/rpc/api.go
@@ -1,28 +1,28 @@
-package reputation
+package rpc
 
 import (
 	"context"
 
 	ma "github.com/multiformats/go-multiaddr"
-	pb "github.com/textileio/powergate/reputation/pb"
+	"github.com/textileio/powergate/reputation"
 )
 
 // Service implements the gprc service
 type Service struct {
-	pb.UnimplementedAPIServer
+	UnimplementedAPIServer
 
-	module *Module
+	module *reputation.Module
 }
 
 // NewService is a helper to create a new Service
-func NewService(m *Module) *Service {
+func NewService(m *reputation.Module) *Service {
 	return &Service{
 		module: m,
 	}
 }
 
 // AddSource calls Module.AddSource
-func (s *Service) AddSource(ctx context.Context, req *pb.AddSourceRequest) (*pb.AddSourceReply, error) {
+func (s *Service) AddSource(ctx context.Context, req *AddSourceRequest) (*AddSourceReply, error) {
 	maddr, err := ma.NewMultiaddr(req.GetMaddr())
 	if err != nil {
 		return nil, err
@@ -30,21 +30,21 @@ func (s *Service) AddSource(ctx context.Context, req *pb.AddSourceRequest) (*pb.
 	if err = s.module.AddSource(req.GetId(), maddr); err != nil {
 		return nil, err
 	}
-	return &pb.AddSourceReply{}, nil
+	return &AddSourceReply{}, nil
 }
 
 // GetTopMiners calls Module.GetTopMiners
-func (s *Service) GetTopMiners(ctx context.Context, req *pb.GetTopMinersRequest) (*pb.GetTopMinersReply, error) {
+func (s *Service) GetTopMiners(ctx context.Context, req *GetTopMinersRequest) (*GetTopMinersReply, error) {
 	minerScores, err := s.module.GetTopMiners(int(req.GetLimit()))
 	if err != nil {
 		return nil, err
 	}
-	pbMinerScores := make([]*pb.MinerScore, len(minerScores))
+	pbMinerScores := make([]*MinerScore, len(minerScores))
 	for i, minerScore := range minerScores {
-		pbMinerScores[i] = &pb.MinerScore{
+		pbMinerScores[i] = &MinerScore{
 			Addr:  minerScore.Addr,
 			Score: int32(minerScore.Score),
 		}
 	}
-	return &pb.GetTopMinersReply{TopMiners: pbMinerScores}, nil
+	return &GetTopMinersReply{TopMiners: pbMinerScores}, nil
 }
